Update attribute type

+++ src/main/java/br/dev/multicode/mcbeers/api/http/requests/BeerRequest.java
@@ -18,7 +18,7 @@ public class BeerRequest {
   private String name;
 
   @NotNull
-  private float alcoholContent;
+  private Double alcoholContent;
 
   @NotBlank
   private String ingredients;
